# backend/src/protocol/rest/error.rs
use serde::Serialize;
use utoipa::ToSchema;

#[derive(Debug, Clone, Copy, Eq, PartialEq, serde::Deserialize, Serialize, ToSchema)]
#[serde(rename_all = "snake_case")]
pub enum ApiErrorCode {
    AuthCredentialsInvalid,
    AuthCredentialsRequired,
    AuthDisplayNameInvalid,
    AuthEmailConflict,
    AuthEmailInvalid,
    AuthEmailRequired,
    AuthLocalLoginDisabled,
    AuthLocalRegistrationDisabled,
    AuthLastLoginMethod,
    AuthPasswordRequired,
    AuthPasswordTooShort,
    AuthPersonalAccessTokenRefused,
    AuthProviderAccountConflict,
    AuthRequired,
    AuthServiceUnavailable,
    AuthUsernameConflict,
    AuthUsernameInvalid,
    AuthUsernameRequired,
    AuthorizationUnavailable,
    BadRequest,
    Unauthorized,
    Forbidden,
    NotFound,
    Conflict,
    ClientIncompatible,
    ExternalGitAuthorizationRequired,
    ExternalGitConnectionInUse,
    ExternalGitNotConfigured,
    ExternalGitOperationConflict,
    ExternalGitProviderUnavailable,
    ExternalGitRepositoryConflict,
    ExternalGitRepositoryForbidden,
    ExternalGitRepositoryNotFound,
    PayloadTooLarge,
    PdfArtifactInvalid,
    PdfArtifactNotFound,
    PdfArtifactTooLarge,
    ProjectAccessForbidden,
    ProjectAssetInvalid,
    ProjectAssetNotFound,
    ProjectAssetTooLarge,
    ProjectContentChanged,
    ProjectContentEpochRequired,
    ProjectDescriptionInvalid,
    ProjectDocumentChanged,
    ProjectDocumentNotFound,
    ProjectNameInvalid,
    ProjectNotFound,
    ProjectPathConflict,
    ProjectPathInvalid,
    ProjectServiceUnavailable,
    ProjectThumbnailInvalid,
    ProjectThumbnailNotFound,
    ProjectThumbnailUnavailable,
    ProjectTypeDisabled,
    RevisionServiceUnavailable,
    RevisionNotFound,
    RevisionSummaryInvalid,
    UnprocessableEntity,
    PreconditionRequired,
    ProcessingArtifactNotFound,
    ProcessingIdempotencyConflict,
    ProcessingIdempotencyInvalid,
    ProcessingIdempotencyRequired,
    ProcessingInputInvalid,
    ProcessingInputTooLarge,
    ProcessingJobNotFound,
    ProcessingOperationInvalid,
    ProcessingQueueFull,
    ProcessingServiceUnavailable,
    ProcessingUnavailable,
    TooManyRequests,
    BadGateway,
    ServiceUnavailable,
    SiteAdminRequired,
    TemplateAssetTooLarge,
    TemplateGalleryUnavailable,
    TemplateInstantiationUnavailable,
    TemplateNotFound,
    TemplateOrganizationGrantNotFound,
    TemplateOrganizationMembershipRequired,
    TemplateRequired,
    TemplateServiceUnavailable,
    TemplateThumbnailNotFound,
    TemplateThumbnailUnavailable,
    InternalError,
    RequestFailed,
}

#[derive(Debug, Serialize, ToSchema)]
pub struct ApiErrorResponse {
    pub code: ApiErrorCode,
    pub message: String,
    #[serde(skip_serializing_if = "Option::is_none")]
    #[schema(required = false)]
    pub request_id: Option<String>,
}