# protocol/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TOSS API",
    "description": "Versioned browser-to-server protocol for the TOSS modular monolith.",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/projects": {
      "get": {
        "tags": [
          "workspace"
        ],
        "operationId": "list_projects",
        "parameters": [
          {
            "name": "include_archived",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectListResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "workspace"
        ],
        "operationId": "create_project",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorCode": {
        "type": "string",
        "enum": [
          "auth_credentials_invalid",
          "auth_credentials_required",
          "auth_display_name_invalid",
          "auth_email_conflict",
          "auth_email_invalid",
          "auth_email_required",
          "auth_local_login_disabled",
          "auth_local_registration_disabled",
          "auth_last_login_method",
          "auth_password_required",
          "auth_password_too_short",
          "auth_personal_access_token_refused",
          "auth_provider_account_conflict",
          "auth_required",
          "auth_service_unavailable",
          "auth_username_conflict",
          "auth_username_invalid",
          "auth_username_required",
          "authorization_unavailable",
          "bad_request",
          "unauthorized",
          "forbidden",
          "not_found",
          "conflict",
          "client_incompatible",
          "external_git_authorization_required",
          "external_git_connection_in_use",
          "external_git_not_configured",
          "external_git_operation_conflict",
          "external_git_provider_unavailable",
          "external_git_repository_conflict",
          "external_git_repository_forbidden",
          "external_git_repository_not_found",
          "payload_too_large",
          "pdf_artifact_invalid",
          "pdf_artifact_not_found",
          "pdf_artifact_too_large",
          "project_access_forbidden",
          "project_asset_invalid",
          "project_asset_not_found",
          "project_asset_too_large",
          "project_content_changed",
          "project_content_epoch_required",
          "project_description_invalid",
          "project_document_changed",
          "project_document_not_found",
          "project_name_invalid",
          "project_not_found",
          "project_path_conflict",
          "project_path_invalid",
          "project_service_unavailable",
          "project_thumbnail_invalid",
          "project_thumbnail_not_found",
          "project_thumbnail_unavailable",
          "project_type_disabled",
          "revision_service_unavailable",
          "revision_not_found",
          "revision_summary_invalid",
          "unprocessable_entity",
          "precondition_required",
          "processing_artifact_not_found",
          "processing_idempotency_conflict",
          "processing_idempotency_invalid",
          "processing_idempotency_required",
          "processing_input_invalid",
          "processing_input_too_large",
          "processing_job_not_found",
          "processing_operation_invalid",
          "processing_queue_full",
          "processing_service_unavailable",
          "processing_unavailable",
          "too_many_requests",
          "bad_gateway",
          "service_unavailable",
          "site_admin_required",
          "template_asset_too_large",
          "template_gallery_unavailable",
          "template_instantiation_unavailable",
          "template_not_found",
          "template_organization_grant_not_found",
          "template_organization_membership_required",
          "template_required",
          "template_service_unavailable",
          "template_thumbnail_not_found",
          "template_thumbnail_unavailable",
          "internal_error",
          "request_failed"
        ]
      },
      "ApiErrorResponse": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ApiErrorCode"
          },
          "message": {
            "type": "string"
          },
          "request_id": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "CreateProjectInput": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "project_type": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ProjectType"
              }
            ]
          },
          "latex_engine": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/LatexEngine"
              }
            ]
          }
        }
      },
...
      "LatexEngine": {
        "type": "string",
        "enum": [
          "pdftex",
          "xetex"
        ]
      },
...
      "Project": {
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "project_type",
          "latex_engine",
          "owner_user_id",
          "owner_display_name",
          "my_role",
          "can_read",
          "is_template",
          "has_thumbnail",
          "created_at",
          "last_edited_at",
          "archived",
          "archived_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "project_type": {
            "$ref": "#/components/schemas/ProjectType"
          },
          "latex_engine": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/LatexEngine"
              }
            ]
          },
          "owner_user_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "owner_display_name": {
            "type": "string"
          },
          "my_role": {
            "$ref": "#/components/schemas/ProjectRole"
          },
          "can_read": {
            "type": "boolean"
          },
          "is_template": {
            "type": "boolean"
          },
          "has_thumbnail": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "last_edited_at": {
            "type": "string",
            "format": "date-time"
          },
          "archived": {
            "type": "boolean"
          },
          "archived_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
...
      "ProjectListResponse": {
        "type": "object",
        "required": [
          "projects"
        ],
        "properties": {
          "projects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Project"
            }
          }
        }
      },
...
      "ProjectRole": {
        "type": "string",
        "enum": [
          "Owner",
          "ReadWrite",
          "ReadOnly"
        ]
      },
...
      "ProjectType": {
        "type": "string",
        "enum": [
          "typst",
          "latex"
        ]
      },
...
    }
  },
...
}